+++ Samples~/Scripts/ExpandableExample.cs
@@ -5,5 +5,7 @@ namespace SaintsField.Samples.Scripts
     public class ExpandableExample : MonoBehaviour
     {
         [SerializeField, Expandable] private Scriptable _scriptable;
+
+        [SerializeField, Expandable] private DropdownExample _dropdownExample;
     }
 }