+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.4 ##
 
 *   Add `AddressableLabel`. This only works if you have `Addressable` installed in your project.
+*   Add `AddressableAddress`
 
 ## 1.1.3 ##