+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 *   Add `AddressableLabel`. This only works if you have `Addressable` installed in your project.
 *   Add `AddressableAddress`. This only works if you have `Addressable` installed in your project.
+*   Add `GetScriptableObject`
 
 ## 1.1.3 ##